+++ GEMValidation/interface/ChamberMatcher.h
@@ -8,7 +8,7 @@
  Original Author:  "Sven Dildick"
 */
 
-#include "GEMCode/GEMValidation/src/BaseMatcher.h"
+#include "GEMCode/GEMValidation/interface/BaseMatcher.h"
 
 #include "SimDataFormats/Track/interface/SimTrackContainer.h"
 #include "SimDataFormats/Vertex/interface/SimVertexContainer.h"
