+++ GEMValidation/plugins/MuonSimHitAnalyzer.cc
@@ -30,7 +30,7 @@
 
 #include "CommonTools/UtilAlgos/interface/TFileService.h"
 
-#include "GEMCode/GEMValidation/src/SimTrackMatchManager.h"
+#include "GEMCode/GEMValidation/interface/SimTrackMatchManager.h"
 
 #include "TTree.h"
