+++ GEMValidation/src/RPCDigiMatcher.cc
@@ -1,5 +1,5 @@
-#include "RPCDigiMatcher.h"
-#include "SimHitMatcher.h"
+#include "GEMCode/GEMValidation/interface/RPCDigiMatcher.h"
+#include "GEMCode/GEMValidation/interface/SimHitMatcher.h"
 
 #include "DataFormats/MuonDetId/interface/CSCDetId.h"
 #include "DataFormats/MuonDetId/interface/CSCTriggerNumbering.h"
