+++ GEMValidation/src/SimHitMatcher.cc
@@ -1,4 +1,4 @@
-#include "GEMCode/GEMValidation/src/SimHitMatcher.h"
+#include "GEMCode/GEMValidation/interface/SimHitMatcher.h"
 
 #include <algorithm>
 #include <iomanip>
