+++ GEMValidation/src/SimTrackMatchManager.cc
@@ -1,4 +1,4 @@
-#include "GEMCode/GEMValidation/src/SimTrackMatchManager.h"
+#include "GEMCode/GEMValidation/interface/SimTrackMatchManager.h"
 
 SimTrackMatchManager::SimTrackMatchManager(const SimTrack& t, const SimVertex& v,
       const edm::ParameterSet& ps, const edm::Event& ev, const edm::EventSetup& es)
