+++ GEMValidation/src/TFCand.cc
@@ -1,4 +1,4 @@
-#include "GEMCode/GEMValidation/src/TFCand.h"
+#include "GEMCode/GEMValidation/interface/TFCand.h"
 
 TFCand::TFCand(const L1MuRegionalCand* t)
 {
